requirements.txt
altgraph==0.17
antiorm==1.2.1
appdirs==1.4.4
APScheduler==3.6.3
astroid==2.4.0
attrs==20.3.0
autobahn==21.2.1
Automat==20.2.0
autopep8==1.5.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.3
constantly==15.1.0
cryptography==3.4.4
dateparser==1.0.0
db==0.1.1
db-sqlite3==0.0.1
Flask==1.1.2
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.1
future==0.18.2
hyperlink==21.0.0
idna==2.10
incremental==17.5.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mpmath==1.1.0
numpy==1.19.5
pefile==2019.4.18
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pygame==1.9.6
PyHamcrest==2.0.2
pyinstaller==4.2
pyinstaller-hooks-contrib==2020.11
pylint==2.5.0
pyOpenSSL==20.0.1
PyQt5==5.14.2
PyQt5-sip==12.7.2
python-binance==0.7.9
python-dateutil==2.8.1
python-telegram-bot==13.2
pytz==2021.1
pywin32-ctypes==0.2.0
regex==2020.11.13
requests==2.25.1
schedule==1.0.0
service-identity==18.1.0
six==1.14.0
SQLAlchemy==1.3.17
sympy==1.7.1
telegram-send==0.25
toml==0.10.0
tornado==6.1
Twisted==20.3.0
txaio==20.12.1
tzlocal==2.1
ujson==4.0.2
urllib3==1.26.3
Werkzeug==1.0.1
wrapt==1.12.1
zope.interface==5.2.0
